requirements.txt
pandas==2.1.4
requests==2.31.0
web3==6.12.0
numpy==1.24.3
scikit-learn==1.3.2
openpyxl==3.1.2
python-dotenv==1.0.0